requirements.txt
aniso8601==7.0.0
backcall==0.2.0
decorator==4.4.2
Django==2.1.4
django-environ==0.4.5
django-filter==2.0.0
django-graphql-jwt==0.2.1
graphene==2.1.8
graphene-django==2.2.0
graphql-core==2.3.2
graphql-relay==2.0.1
ipdb==0.13.3
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
traitlets==4.3.3
wcwidth==0.2.5
